Week3-Assignment 3: Collinear Points
[Problem]
Given a set of n distinct points in the plane, find every (maximal)
line segment that connects a subset of 4 or more of the points.

[Solution]
Add two different compare way to Point.java, one is by comparable interface
and the other by comparator. And implement two different methods to find
the collinear points. The first one is brute froce way with O(n^4) time
complexity and the other one is based on quick sort whose time complexity is
O(n^2*logn).

[Tests]
Passed all the tests from the course grading system and got 100/100.